pyproject.toml
[project]
name = "vedic-ai-framework"
version = "0.1.0"
description = "AI-powered Vedic astrology computation and interpretation"
requires-python = ">=3.10"
license = "MIT"
authors = [{name = "Manish Chaurasia"}]
dependencies = [
    "pyswisseph>=2.10.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "geopy>=2.4.0",
    "timezonefinder>=6.0",
    "pytz>=2024.1",
    "jinja2>=3.0",
    "rich>=13.0",
]

[project.optional-dependencies]
ollama = ["ollama>=0.4.0"]
groq = ["groq>=0.9.0"]
claude = ["anthropic>=0.39.0"]
openai = ["openai>=1.50.0"]
pdf = ["reportlab>=4.0"]
telegram = ["python-telegram-bot>=21.0"]
whisper = ["openai-whisper>=20231117"]
all = ["vedic-ai-framework[ollama,groq,claude,openai,pdf,telegram]"]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[project.scripts]
jyotish = "jyotish.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["jyotish*"]
